// apps/web/src/modules/user-profile/domain/types/user-account-sidebar-view.ts
export type UserAccountSidebarSocialLink = {
  type: string;
  value: string;
  href: string;
};

export type UserAccountSidebarCryptoWallet = {
  id: string;
  label: string;
  shortName: string;
  abbreviation: string;
  address: string;
  icon: string;
  coingeckoId: string;
};

export type UserAccountSidebarView = {
  about: string;
  location: string | null;
  website: string | null;
  email: string | null;
  joinedAt: string | null;
  expertiseWeight: number;
  lastActivityAt: string | null;
  totalVoteValueUsd: number;
  socialLinks: UserAccountSidebarSocialLink[];
  cryptoWallets: UserAccountSidebarCryptoWallet[];
  waiv: {
    upvotingManaPercent: number;
    downvotingManaPercent: number;
    voteValueUsd: number;
  };
  hive: {
    reputation: number;
    upvotingManaPercent: number;
    downvotingManaPercent: number;
    resourceCreditsPercent: number;
    voteValueUsd: number;
  };
};